was able to refactor the initial version history, constructed a page for the version info
