+++ SpillTracker/SpillTracker/Models/ListOfCommitsVM.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace SpillTracker.Models
+{
+    public class ListOfCommitsVM
+    {
+        public List<Commit> commits = new List<Commit>();
+    }
+}
