+++ SpillTracker/SpillTracker/SpillTracker.csproj
@@ -58,6 +58,7 @@
   <ItemGroup>
     <None Include="Views\Facilities\EditFacChems.cshtml" />
     <None Include="Views\Home\Guide.cshtml" />
+    <None Include="Views\VersionHistory\Index.cshtml" />
 
     <None Include="wwwroot\js\adminpanel.js" />
     <None Include="wwwroot\js\adminpanel2.js" />
