+++ SpillTracker/SpillTracker/Views/Shared/_Layout.cshtml
@@ -86,9 +86,8 @@
         <div class="container">
             &copy; 2021 - SpillTracker - NATR SS
 
-            <button style="margin: auto;" type="button" class="btn btn-secondary dropdown-toggle" data-toggle="dropdown" aria-haspopup="true" aria-expanded="true">
-                Version History
-            </button>
+            <a class="nav-link text-light" asp-controller="VersionHistory" asp-action="Index">Version History</a>
+            
             <div id="versionH" class="dropdown-menu" style="width: 90vw; height: 20rem; overflow-y: auto;">
 
             </div>
